+++ examples/storybook/stories/Components/EditorWrapper.tsx
@@ -310,6 +310,7 @@ interface Props {
   rcProps?: Record<string, any>;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   experiments?: Record<string, any>;
+  modalSettings?: { container: HTMLElement };
 }
 
 class EditorWrapper extends React.Component<Props> {
@@ -339,6 +340,7 @@ class EditorWrapper extends React.Component<Props> {
       onFocus,
       rcProps = {},
       experiments,
+      modalSettings,
     } = this.props;
 
     return (
@@ -360,6 +362,7 @@ class EditorWrapper extends React.Component<Props> {
           _rcProps={rcProps}
           onAtomicBlockFocus={d => console.log('onAtomicBlockFocus', d)} // eslint-disable-line
           commands={commands}
+          modalSettings={modalSettings}
         >
           <RichContentEditor
             onFocus={onFocus}