+++ CHANGELOG.md
@@ -27,7 +27,8 @@
   - [#3878](https://github.com/wix/ricos/pull/3878) fix mentions selected / hovered background
 - `settings-modals`
   - [#3867](https://github.com/wix/ricos/pull/3867) relate experiments & getComponentData as optional
-- `ricos-eitor`
+- `ricos-editor`
+  - [#3885](https://github.com/wix/ricos/pull/3885) fix modal scoping issue /use RicosPortal in RicosModal
   - [#3886](https://github.com/wix/ricos/pull/3886) fix flaky plus button position on tiptap
 
 ### :house: Internal